attribute test fixtures relocated
