+++ tests/Attribute/BuilderTest.php
@@ -9,31 +9,31 @@ use Membrane\Attribute\ClassWithAttributes;
 use Membrane\Builder\Specification;
 use Membrane\Exception\CannotProcessProperty;
 use Membrane\Filter\CreateObject\WithNamedArguments;
-use Membrane\Fixtures\ArraySumFilter;
-use Membrane\Fixtures\ClassThatOverridesProcessorType;
-use Membrane\Fixtures\ClassWithClassArrayPropertyIsIntValidator;
-use Membrane\Fixtures\ClassWithClassProperty;
-use Membrane\Fixtures\ClassWithCompoundPropertyType;
-use Membrane\Fixtures\ClassWithDateTimeProperty;
-use Membrane\Fixtures\ClassWithIntArrayPropertyBeforeSet;
-use Membrane\Fixtures\ClassWithIntArrayPropertyIsIntValidator;
-use Membrane\Fixtures\ClassWithIntProperty;
-use Membrane\Fixtures\ClassWithIntPropertyIgnoredProperty;
-use Membrane\Fixtures\ClassWithIntPropertyIsIntValidator;
-use Membrane\Fixtures\ClassWithNestedCollection;
-use Membrane\Fixtures\ClassWithNoSubTypeHint;
-use Membrane\Fixtures\ClassWithNoTypeHint;
-use Membrane\Fixtures\ClassWithPromotedPropertyAfterSet;
-use Membrane\Fixtures\ClassWithStringPropertyBeforeSet;
-use Membrane\Fixtures\Docs\BlogPostFromNamedArguments;
-use Membrane\Fixtures\Docs\BlogPostIsItAString;
-use Membrane\Fixtures\Docs\BlogPostMakeItAString;
-use Membrane\Fixtures\Docs\BlogPostMaxTags;
-use Membrane\Fixtures\Docs\BlogPostRegexAndMaxLength;
-use Membrane\Fixtures\Docs\BlogPostRequiredFields;
-use Membrane\Fixtures\Docs\BlogPostWithAllOf;
-use Membrane\Fixtures\EmptyClass;
-use Membrane\Fixtures\EmptyClassWithIgnoredProperty;
+use Membrane\Fixtures\Attribute\ArraySumFilter;
+use Membrane\Fixtures\Attribute\ClassThatOverridesProcessorType;
+use Membrane\Fixtures\Attribute\ClassWithClassArrayPropertyIsIntValidator;
+use Membrane\Fixtures\Attribute\ClassWithClassProperty;
+use Membrane\Fixtures\Attribute\ClassWithCompoundPropertyType;
+use Membrane\Fixtures\Attribute\ClassWithDateTimeProperty;
+use Membrane\Fixtures\Attribute\ClassWithIntArrayPropertyBeforeSet;
+use Membrane\Fixtures\Attribute\ClassWithIntArrayPropertyIsIntValidator;
+use Membrane\Fixtures\Attribute\ClassWithIntProperty;
+use Membrane\Fixtures\Attribute\ClassWithIntPropertyIgnoredProperty;
+use Membrane\Fixtures\Attribute\ClassWithIntPropertyIsIntValidator;
+use Membrane\Fixtures\Attribute\ClassWithNestedCollection;
+use Membrane\Fixtures\Attribute\ClassWithNoSubTypeHint;
+use Membrane\Fixtures\Attribute\ClassWithNoTypeHint;
+use Membrane\Fixtures\Attribute\ClassWithPromotedPropertyAfterSet;
+use Membrane\Fixtures\Attribute\ClassWithStringPropertyBeforeSet;
+use Membrane\Fixtures\Attribute\Docs\BlogPostFromNamedArguments;
+use Membrane\Fixtures\Attribute\Docs\BlogPostIsItAString;
+use Membrane\Fixtures\Attribute\Docs\BlogPostMakeItAString;
+use Membrane\Fixtures\Attribute\Docs\BlogPostMaxTags;
+use Membrane\Fixtures\Attribute\Docs\BlogPostRegexAndMaxLength;
+use Membrane\Fixtures\Attribute\Docs\BlogPostRequiredFields;
+use Membrane\Fixtures\Attribute\Docs\BlogPostWithAllOf;
+use Membrane\Fixtures\Attribute\EmptyClass;
+use Membrane\Fixtures\Attribute\EmptyClassWithIgnoredProperty;
 use Membrane\Processor\AfterSet;
 use Membrane\Processor\BeforeSet;
 use Membrane\Processor\Collection;
