+++ tests/fixtures/Attribute/ClassWithStringPropertyBeforeSet.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Membrane\Fixtures;
+namespace Membrane\Fixtures\Attribute;
 
 use Membrane\Attribute\Placement;
 use Membrane\Attribute\SetFilterOrValidator;
